+++ packages/svelte/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Next
 
+* Add `hash`, `query`, and `state` props to `<Link>`, remove `details`.
 * Add `curiStore` function to automatically sync router and store. This can create a new store or update an existing store.
 
 ## 1.0.0-beta.4